backend/requirements.txt
# © 2026 Octávio Filipe Gonçalves
# Callsign: CT7BFV
# License: GNU AGPL-3.0 (https://www.gnu.org/licenses/agpl-3.0.html)
# Last update: 2026-02-22 00:34:50 UTC

fastapi
uvicorn
numpy
scipy
pytest
psutil
pyyaml
jsonschema
